requirements.txt
google-generativeai
langchain
langchain-community
faiss-cpu
beautifulsoup4
requests
python-dotenv
pypdf